% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{mutation_GDSC}
\alias{mutation_GDSC}
\title{mutation_GDSC}
\format{
an object of "tbl_df" (tidy table)
}
\usage{
data(mutation_GDSC)
}
\description{
This cancer type and genomic background annotation for cancer cell lines, use for the analysis of the GDSC dataset in the package vignetee.
}
\examples{
data(mutation_GDSC)
}
\keyword{datasets}
