% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{responseInput}
\alias{responseInput}
\title{responseInput}
\format{
an object of matrix
}
\usage{
data(responseInput)
}
\description{
A toy data set that contains processed drug response matrix for examples and test of runLASSORegression function. 
Rows contain drugs and columns contain samples.
}
\examples{
data(responseInput)
}
\keyword{datasets}
