% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depInfeR.R
\name{runLASSORegression}
\alias{runLASSORegression}
\title{Main function to run LASSO regression}
\usage{
runLASSORegression(TargetMatrix, ResponseMatrix, cores = 1, repeats = 100)
}
\arguments{
\item{TargetMatrix}{Pre-processed drug-protein affinity matrix. Rows should contain drugs and columns should contain targets.}

\item{ResponseMatrix}{Pre-processed drug-response matrix. Rows should contain drugs and columns should contain samples.}

\item{cores}{A integer variable specifying the number of cores. Multi-core parallelization may only work for Mac OS and Linux.}

\item{repeats}{A integer variable specifying the number of regression repeats.}
}
\value{
Pre-processed drug-protein affinity matrix
}
\description{
Main function to run LASSO regression
}
\examples{
data(responseInput) #load drug response matrix
data(targetInput) #load drug-target affinity matrix
runLASSORegression(TargetMatrix = targetInput, ResponseMatrix = responseInput, repeats = 5)
# Please refer to the package vignette for more detailed information about this function.
# For the mathematical model behind this function, 
# please refer to our preprint on bioRxiv: https://doi.org/10.1101/2022.01.11.475864

}
