% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{targetInput}
\alias{targetInput}
\title{targetInput}
\format{
an object of matrix
}
\usage{
data(targetInput)
}
\description{
A toy data set that contains processed drug-target affinity matrix for examples and test of runLASSOregression function. 
Rows contain drugs and columns contain targets.
}
\examples{
data(targetInput)
}
\keyword{datasets}
