% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{targetMatrix}
\alias{targetMatrix}
\title{targetMatrix}
\format{
an object of matrix
}
\usage{
data(targetMatrix)
}
\description{
Drug-target affinity matrix
}
\examples{
data(targetMatrix)
}
\keyword{datasets}
