% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{drug_response_GDSC}
\alias{drug_response_GDSC}
\title{drug_response_GDSC}
\format{
an object of "tbl_df" (tidy table)
}
\usage{
data(drug_response_GDSC)
}
\description{
This is the Genomics of Drug Sensitivity in Cancer (GDSC) drug sensitivity dataset, downloaded from https://www.cancerrxgene.org/.
}
\examples{
data(drug_response_GDSC)
}
\keyword{datasets}
