% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{responseInput}
\alias{responseInput}
\title{responseInput}
\format{
an object of matrix
}
\usage{
data(responseInput)
}
\description{
Processed drug response matrix for examples and test of runLASSOregression
}
\examples{
data(responseInput)
}
\keyword{datasets}
