% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depInfeR.R
\name{runLASSOregression}
\alias{runLASSOregression}
\title{Main function to run LASSO regression}
\usage{
runLASSOregression(TargetMatrix, ResponseMatrix, cores = 1, repeats = 100)
}
\arguments{
\item{TargetMatrix}{Pre-processed drug-protein affinity matrix. Rows should contain drugs and columns should contain targets.}

\item{ResponseMatrix}{Pre-processed drug-response matrix. Rows should contain drugs and columns should contain samples.}

\item{cores}{A numeric variable specifying the number of cores. Multi-core parallelization may only work for Mac OS and Linux.}

\item{repeats}{A numeric variable specifying the number of regression repeats.}
}
\value{
Pre-processed drug-protein affinity matrix
}
\description{
Main function to run LASSO regression
}
\examples{
data(responseInput) #load drug response matrix
data(targetInput) #load drug-target affinity matrix
runLASSOregression(TargetMatrix = targetInput, ResponseMatrix = responseInput, repeats = 5)
# Please refer to the package vignette for more detailed information about this function.


}
