% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{targetInput}
\alias{targetInput}
\title{targetInput}
\format{
an object of matrix
}
\usage{
data(targetInput)
}
\description{
Processed drug-target affinity matrix for examples and test of runLASSOregression
}
\examples{
data(targetInput)
}
\keyword{datasets}
