% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{targetsGDSC}
\alias{targetsGDSC}
\title{targetsGDSC}
\format{
an object of "tbl_df" (tidy table)
}
\usage{
data(targetsGDSC)
}
\description{
This drug-protein affinity profiling data for the analysis of the GDSC dataset - a subset of the data provided by Klaeger et al. 2017.
}
\examples{
data(targetsGDSC)
}
\keyword{datasets}
